backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
scikit-learn
numpy
pandas
kafka-python
python-dotenv
alembic
pytest